refactor: AuthFacade 추가 및 적용 #130
