+++ src/main/java/org/sopt/app/presentation/auth/AppAuthController.java
@@ -8,28 +8,21 @@ import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import javax.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import lombok.val;
-import org.sopt.app.application.auth.JwtTokenService;
-import org.sopt.app.application.auth.PlaygroundAuthService;
-import org.sopt.app.application.soptamp.SoptampPointService;
-import org.sopt.app.application.soptamp.SoptampUserService;
-import org.sopt.app.application.user.UserService;
+import org.sopt.app.facade.AuthFacade;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.PatchMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/api/v2/auth")
 public class AppAuthController {
 
-    private final PlaygroundAuthService playgroundAuthService;
-    private final UserService userService;
-    private final SoptampUserService soptampUserService;
-    private final SoptampPointService soptampPointService;
-
-    private final JwtTokenService jwtTokenService;
-
-    private final AppAuthResponseMapper authResponseMapper;
+    private final AuthFacade authFacade;
 
     @Operation(summary = "플그로 로그인/회원가입")
     @ApiResponses({
@@ -41,25 +34,7 @@ public class AppAuthController {
     public ResponseEntity<AppAuthResponse.Token> playgroundLogin(
             @Valid @RequestBody AppAuthRequest.CodeRequest codeRequest
     ) {
-        // PlayGround SSO Auth 를 통해 accessToken 받아옴
-        val temporaryToken = playgroundAuthService.getPlaygroundAccessToken(codeRequest);
-        // PlayGround Auth Access Token 받옴
-        val playgroundToken = playgroundAuthService.refreshPlaygroundToken(temporaryToken);
-        // PlayGround User Info 받아옴
-        val playgroundMember = playgroundAuthService.getPlaygroundInfo(playgroundToken.getAccessToken());
-
-        val userId = userService.loginWithUserPlaygroundId(playgroundMember);
-        val soptampUserId = soptampUserService.updateSoptampUser(playgroundMember.getName(), userId.getId());
-        soptampPointService.upsertSoptampPoint(playgroundMember.getStatus(), soptampUserId);
-
-        // Response 할 Body 생성
-        val appToken = jwtTokenService.issueNewTokens(userId, playgroundMember);
-        val response = authResponseMapper.of(
-                appToken.getAccessToken()
-                , appToken.getRefreshToken()
-                , playgroundMember.getAccessToken()
-                , playgroundMember.getStatus()
-        );
+        val response = authFacade.loginWithPlayground(codeRequest);
         return ResponseEntity.status(HttpStatus.OK).body(response);
     }
 
@@ -73,14 +48,7 @@ public class AppAuthController {
     public ResponseEntity<AppAuthResponse.Token> refreshToken(
             @Valid @RequestBody AppAuthRequest.RefreshRequest refreshRequest
     ) {
-        val userId = jwtTokenService.getUserIdFromJwtToken(refreshRequest.getRefreshToken());
-        val existingToken = userService.getPlaygroundToken(userId);
-        val playgroundToken = playgroundAuthService.refreshPlaygroundToken(existingToken);
-        val playgroundMember = playgroundAuthService.getPlaygroundInfo(playgroundToken.getAccessToken());
-
-        val appToken = jwtTokenService.issueNewTokens(userId, playgroundMember);
-        val response = authResponseMapper.of(appToken.getAccessToken(), appToken.getRefreshToken(),
-                playgroundToken.getAccessToken(), playgroundMember.getStatus());
+        val response = authFacade.getRefreshToken(refreshRequest);
         return ResponseEntity.status(HttpStatus.OK).body(response);
     }
 }
